Ajeitando bug do ranking zerao

+++ cli-lib-main/include/tela.h
@@ -3,7 +3,7 @@
 
 typedef struct {
     char nome[21];
-    int score;
+    long long score;
 } Jogador;
 
 // Funções de controle do terminal
